refractor: move mapBootcampToSaltData, mapBootcampToSaltData2 to dataHelpers.ts
