+++ frontend/src/components/Feature/Diplomadataform/DiplomaDataForm.tsx
@@ -3,11 +3,10 @@ import { useForm } from "react-hook-form";
 import { TemplateResponse, SaltData, Student, FormDataUpdateRequest, TrackResponse, BootcampResponse, pdfGenerationResponse } from "../../../util/types";
 import { FileUpload } from "../../MenuItems/Inputs/FileUploader";
 import { ParseFileData } from '../../../services/InputFileService';
-import { mapBootcampToSaltData2 } from "../../../util/helper";
 import './DiplomaDataForm.css';
 import { AlertPopup, CustomAlertPopupProps, PopupType } from "../../MenuItems/Popups/AlertPopup";
 import { Template } from "@pdfme/common";
-import { mapTemplateInputsBootcampsToTemplateViewer, templateInputsFromBootcampData } from "../../../util/dataHelpers";
+import { mapBootcampToSaltData2, mapTemplateInputsBootcampsToTemplateViewer, templateInputsFromBootcampData } from "../../../util/dataHelpers";
 import { SelectOptions } from "../../MenuItems/Inputs/SelectOptions";
 import { CheckboxGroup } from "../../MenuItems/Inputs/CheckBoxGroup";
 import { UpdateIcon } from "../../MenuItems/Icons/UpdateIcon";
