+++ frontend/src/util/dataHelpers.ts
@@ -507,3 +507,25 @@ export const createBlankTemplate = (templateName: string) => {
       }
     return blankTemplate;
 }
+
+
+export function mapBootcampToSaltData(bootcamp: BootcampResponse, template: TemplateResponse ): SaltData {
+  return {
+      guidId: bootcamp.guidId,
+      classname: bootcamp.name ,
+      dategraduate: bootcamp.graduationDate.toString().slice(0, 10),
+      students: bootcamp.students,
+      template: template
+  };
+}
+
+export function mapBootcampToSaltData2(TrackName: string, bootcamp: BootcampResponse, template: TemplateResponse ): SaltData {
+  return {
+      guidId: bootcamp.guidId,
+      classname: bootcamp.name ,
+      dategraduate: bootcamp.graduationDate.toString().slice(0, 10),
+      students: bootcamp.students,
+      template: template,
+      displayName: "Fullstack " + TrackName 
+  };
+}
