+++ frontend/src/util/helper.ts
@@ -97,24 +97,3 @@ export const isJsonString = (str: string) => {
   }
   return true;
 };
-
-export function mapBootcampToSaltData(bootcamp: BootcampResponse, template: TemplateResponse ): SaltData {
-  return {
-      guidId: bootcamp.guidId,
-      classname: bootcamp.name ,
-      dategraduate: bootcamp.graduationDate.toString().slice(0, 10),
-      students: bootcamp.students,
-      template: template
-  };
-}
-
-export function mapBootcampToSaltData2(TrackName: string, bootcamp: BootcampResponse, template: TemplateResponse ): SaltData {
-  return {
-      guidId: bootcamp.guidId,
-      classname: bootcamp.name ,
-      dategraduate: bootcamp.graduationDate.toString().slice(0, 10),
-      students: bootcamp.students,
-      template: template,
-      displayName: "Fullstack " + TrackName 
-  };
-}
